WIP Added checks for ARROW_HDFS, switched to c++ for MSVC where through <ranges> there are int128 types, simplified bazel compilation

+++ cpp/src/arrow/util/basic_decimal.cc
@@ -169,8 +169,8 @@ struct uint128_t {
 
   explicit uint128_t(uint64_t value) : val_(value) {}
 
-  uint64_t hi() { return val_ >> 64; }
-  uint64_t lo() { return val_ & kInt64Mask; }
+  uint64_t hi() { return static_cast<uint64_t>(val_ >> 64); }
+  uint64_t lo() { return static_cast<uint64_t>(val_ & kInt64Mask); }
 
   uint128_t& operator+=(const uint128_t& other) {
     val_ += other.val_;
